Added support if iframe and redirect isn't allowed
There are cases when developing with adaljs where the use of iframes,
redirecting to another page or even changing the the window.location
object is not allowed, for example when developing a google chrome
extension and using the adaljs library.

By adding some extra properties to the initial configuration, the
developer would be able to specify whether the use of iframes and/or
redirect is allowed, as well as a logOutCall function in case the
environment doesn’t allow modifications on the window.location object.